deleted extra component level and transferred tests
