--- web/src/components/Translation/Translation.stories.jsx
@@ -1,17 +1,0 @@
-// Pass props to your component by passing an `args` object to your story
-//
-// ```jsx
-// export const Primary = {
-//  args: {
-//    propName: propValue
-//  }
-// }
-// ```
-//
-// See https://storybook.js.org/docs/react/writing-stories/args.
-
-import Translation from './Translation'
-
-export default { component: Translation }
-
-export const Primary = {}
